Add multiple survey selection functionality
