+++ src/UI/NextPageButtonLink.tsx
@@ -4,10 +4,11 @@ interface NextPageButtonLinkProps{
   pageName: string;
   msg: string;
   disabled?: boolean;
+  query? : {[key: string]: string}
 }
 
-export const NextPageButtonLink = ({pageName, msg, disabled}: NextPageButtonLinkProps) => {
-  const href = `./${pageName}`;
+export const NextPageButtonLink = ({pageName, msg, disabled, query}: NextPageButtonLinkProps) => {
+  const href = {pathname: `./${pageName}`, query: query};
   return (
   <div>
     <Link href={href}><button disabled={disabled} className="rounded-full bg-white/10 px-10 py-3 hover:bg-white/20 my-10 text-white">
